backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-socketio==5.10.0
python-engineio==4.8.0
aiohttp==3.9.1
python-dotenv==1.0.0
httpx==0.27.0 